rename package to gitapi to avoid package name conflicts
